// JobTracker/UserInterface/app/types/jobs/create-job-tracker-request.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { Tag } from "../tag/tag";

export interface CreateJobTrackerRequest {
    Keyword: string;
    Source: string;
    Location: string;
    IsActive: boolean;
    Tags: Tag[];
    LastCheckedAt: Date;
    CheckIntervalHours: number;
}